+++ src/main/java/logion/backend/chain/view/Extrinsic.java
@@ -1,6 +1,7 @@
 package logion.backend.chain.view;
 
 import com.fasterxml.jackson.databind.JsonNode;
+import java.math.BigInteger;
 import logion.backend.annotation.View;
 import lombok.Data;
 
@@ -11,7 +12,7 @@ public class Extrinsic {
     Method method;
     Signature signature;
     JsonNode args;
-    Long tip;
+    BigInteger tip;
     Info info;
     Event[] events;
     boolean success;
