+++ src/test/java/logion/backend/model/transaction/TransactionTest.java
@@ -1,5 +1,6 @@
 package logion.backend.model.transaction;
 
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 import logion.backend.model.Ss58Address;
 import logion.backend.model.transaction.Transaction.TransactionId;
@@ -28,7 +29,7 @@ class TransactionTest {
         transaction.from = new Ss58Address("from");
         transaction.to = new Ss58Address("to");
         transaction.createdOn = LocalDateTime.now();
-        transaction.transferValue = 123456;
+        transaction.transferValue = BigInteger.valueOf(123456L);
         return transaction;
     }
 }
